+++ WMS/Controllers/BaseController.cs
@@ -431,7 +431,7 @@ namespace WMS.Controllers
         public void VerifyOrderStatus(int orderId)
         {
             var order = OrderService.GetOrderById(orderId);
-            if (order.OrderStatusID == (int)OrderStatusEnum.Active)
+            if (order.OrderStatusID == OrderStatusEnum.Active)
             {
                 ViewBag.PreventProcessing = false;
